1.Advance_Chatbot_With_Memory/requirements.txt
langchain==0.2.10
langchain-openai==0.1.20
python-dotenv==1.0.1
langchain-community==0.2.10
langchain-groq==0.1.6